- Moved listener and command auto-registration logic to platform module
