+++ springlify-platform/springlify-platform-bukkit/springlify-platform-bukkit.gradle
@@ -1,6 +1,9 @@
 dependencies {
-    compileOnlyApi "com.mojang:authlib:$authlibVersion"
-    api "de.tr7zw:item-nbt-api:$nbtApiVersion"
+    compileOnly "com.destroystokyo.paper:paper-api:$paperVersion"
+    compileOnly "com.mojang:authlib:$authlibVersion"
+
+    implementation "de.tr7zw:item-nbt-api:$nbtApiVersion"
+
     api project(":springlify-starter:springlify-starter-bukkit")
     api project(":springlify-platform:springlify-platform-commons")
