+++ springlify-platform/springlify-platform-bukkit/src/main/java/dev/temez/springlify/platform/annotation/BukkitCommand.java
@@ -1,4 +1,4 @@
-package dev.temez.springlify.starter.annotation;
+package dev.temez.springlify.platform.annotation;
 
 import org.jetbrains.annotations.NotNull;
 import org.springframework.core.annotation.AliasFor;
