+++ springlify-platform/springlify-platform-bukkit/src/main/java/dev/temez/springlify/platform/initializer/EventHandlerInitializer.java
@@ -1,6 +1,6 @@
-package dev.temez.springlify.starter.initializer;
+package dev.temez.springlify.platform.initializer;
 
-import dev.temez.springlify.starter.server.ServerPlatformAdapter;
+import dev.temez.springlify.platform.server.ServerPlatformAdapter;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
 import org.bukkit.event.Listener;
