+++ springlify-platform/springlify-platform-bukkit/src/main/java/dev/temez/springlify/platform/server/BukkitServerPlatformAdapter.java
@@ -1,4 +1,4 @@
-package dev.temez.springlify.starter.server;
+package dev.temez.springlify.platform.server;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
