+++ springlify-starter/springlify-starter-bukkit/springlify-starter-bukkit.gradle
@@ -1,5 +1,6 @@
 dependencies {
-    compileOnlyApi "com.destroystokyo.paper:paper-api:$paperVersion"
+    compileOnly "com.destroystokyo.paper:paper-api:$paperVersion"
+
     api project(":springlify-starter:springlify-starter-commons")
 
     testImplementation "com.destroystokyo.paper:paper-api:$paperVersion"
