+++ springlify-starter/springlify-starter-commons/README.md
@@ -1,4 +1,5 @@
 # springlify-starter-commons
 
-A cross-platform part of `Springlify` starter. 
-It contains necessary abstractions and mechanisms to ensure compatibility with platforms such as `Bukkit`, `Velocity` and others.
+A cross-platform part of `Springlify` starter.
+It contains necessary abstractions and mechanisms to ensure compatibility with platforms such as `Bukkit`, `Velocity`
+and others.
