+++ springlify-starter/springlify-starter-commons/springlify-starter-commons.gradle
@@ -1,4 +1,3 @@
 dependencies {
-    api 'org.springframework.boot:spring-boot-starter'
     api "org.reflections:reflections:$reflectionsApiVersion"
 }
