+++ springlify-starter/springlify-starter-commons/src/main/java/dev/temez/springlify/starter/configuration/SpringlifyPluginConfiguration.java
@@ -2,7 +2,6 @@ package dev.temez.springlify.starter.configuration;
 
 import dev.temez.springlify.starter.initializer.SpringlifyInitializer;
 import dev.temez.springlify.starter.plugin.SpringlifyPlugin;
-import dev.temez.springlify.starter.server.ServerPlatformAdapter;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,10 +9,9 @@ import org.springframework.context.annotation.Configuration;
 /**
  * Spring configuration class for integrating Springlify plugins.
  *
- * <p>This class provides Spring beans for the {@link ServerPlatformAdapter} and {@link SpringlifyInitializer}
+ * <p>This class provides Spring beans for the {@link SpringlifyInitializer}
  * by retrieving them from the provided {@link SpringlifyPlugin} instance.</p>
  *
- * @see ServerPlatformAdapter
  * @see SpringlifyInitializer
  * @see SpringlifyPlugin
  * @since 0.7.0.0-RC1
@@ -21,17 +19,6 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class SpringlifyPluginConfiguration {
 
-  /**
-   * Provides the {@link ServerPlatformAdapter} as a Spring bean.
-   *
-   * @param plugin the Springlify plugin instance
-   * @return the server platform adapter
-   */
-  @Bean
-  @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
-  ServerPlatformAdapter serverPlatformAdapter(@NotNull SpringlifyPlugin plugin) {
-    return plugin.getServerPlatformAdapter();
-  }
 
   /**
    * Provides the {@link SpringlifyInitializer} as a Spring bean.
